chat migrated to app with wearable
