+++ apps/AppWithWearable/lib/custom_code/actions/index.dart
@@ -7,4 +7,3 @@ export 'ble0receive_data.dart' show ble0receiveData;
 export 'ble0send_data.dart' show ble0sendData;
 export 'ble_receive_w_a_v.dart' show bleReceiveWAV;
 export 'ble_receive_wav_test.dart' show bleReceiveWavTest;
-export 'stream_api_response.dart' show streamApiResponse;
