+++ apps/AppWithWearable/lib/index.dart
@@ -1,15 +1,10 @@
-// Export pages
-export '/chat_page/chat_page_widget.dart' show ChatPageWidget;
+export '/pages/chat_page/chat_widget.dart' show ChatPageWidget;
 export '/pages/home_page/home_page_widget.dart' show HomePageWidget;
-// export 'package:friend_private/home_page/home_page_widget.dart' show HomePageWidget;
-export '/onboarding/permission_page/permission_page_widget.dart'
-    show PermissionPageWidget;
+
+export '/onboarding/permission_page/permission_page_widget.dart' show PermissionPageWidget;
 export '/onboarding/welcome/welcome_widget.dart' show WelcomeWidget;
 export '/pages/ble/connected/connected_widget.dart' show ConnectedWidget;
 export '/pages/ble/connecting/connecting_widget.dart' show ConnectingWidget;
-export '/pages/ble/find_devices/find_devices_widget.dart'
-    show FindDevicesWidget;
-export '/pages/ble/scan_devices/scan_devices_widget.dart'
-    show ScanDevicesWidget;
-export '/pages/connect_device/connect_device_widget.dart'
-    show ConnectDeviceWidget;
+export '/pages/ble/find_devices/find_devices_widget.dart' show FindDevicesWidget;
+export '/pages/ble/scan_devices/scan_devices_widget.dart' show ScanDevicesWidget;
+export '/pages/connect_device/connect_device_widget.dart' show ConnectDeviceWidget;
