+++ apps/AppWithWearable/lib/pages/chat_page/chat_model.dart
@@ -1,20 +1,34 @@
+import '/flutter_flow/flutter_flow_util.dart';
+import 'chat_widget.dart' show ChatPageWidget;
 import 'package:flutter/material.dart';
 
-import '/flutter_flow/flutter_flow_util.dart';
-import 'chat_page_widget.dart' show ChatPageWidget;
+class ChatModel extends FlutterFlowModel<ChatPageWidget> {
+  ///  Local state fields for this page.
 
-class ChatPageModel extends FlutterFlowModel<ChatPageWidget> {
+  bool showCommandAlertBool = false;
+
   ///  State fields for stateful widgets in this page.
 
   final unfocusNode = FocusNode();
+
   // State field(s) for Column widget.
   ScrollController? columnController;
+
   // State field(s) for ListView widget.
   ScrollController? listViewController;
+
   // State field(s) for TextField widget.
   FocusNode? textFieldFocusNode;
   TextEditingController? textController;
   String? Function(BuildContext, String?)? textControllerValidator;
+
+  // Stores action output result for [Backend Call - API (Vectorize)] action in IconButton widget.
+  List<double>? vector;
+
+  // Stores action output result for [Backend Call - API (QueryVectors)] action in IconButton widget.
+  List? simillarVectors;
+
+  /// Initialization and disposal methods.
 
   @override
   void initState(BuildContext context) {
@@ -29,5 +43,13 @@ class ChatPageModel extends FlutterFlowModel<ChatPageWidget> {
     listViewController?.dispose();
     textFieldFocusNode?.dispose();
     textController?.dispose();
+  }
+
+  Future pageCommand(BuildContext context) async {}
+
+  Future showCommandAlert(BuildContext context) async {
+    showCommandAlertBool = true;
+    await Future.delayed(const Duration(milliseconds: 5000));
+    showCommandAlertBool = false;
   }
 }
